+++ CHANGELOG.md
@@ -7,10 +7,12 @@
 - add features and tests for author search
 - add features and tests for title search
 - add features and tests for subjects search
+- Add government doc filter for search results and advanced search
+- Add "Clear Filters" button for search
 - Upgrade to NYPL Design System version 1.1.0
 - Upgrade to React 17.0.2
 - Upgrade to Next 12.2.5
-- Add "Clear Filters" button for search
+- Fix undefined-heading id on details list heading bug
 
 ## [0.14.1]